+++ src/Commands/ImportAccountAudience.php
@@ -6,6 +6,7 @@ use Illuminate\Console\Command;
 use OneMediaLabs\MixpostMcp\Concerns\AccountsOption;
 use OneMediaLabs\MixpostMcp\SocialProviders\Mastodon\Jobs\ImportMastodonFollowersJob;
 use OneMediaLabs\MixpostMcp\SocialProviders\Meta\Jobs\ImportFacebookPageFollowersJob;
+use OneMediaLabs\MixpostMcp\SocialProviders\Meta\Jobs\ImportInstagramFollowersJob;
 use OneMediaLabs\MixpostMcp\SocialProviders\Twitter\Jobs\ImportTwitterFollowersJob;
 
 class ImportAccountAudience extends Command
@@ -22,6 +23,7 @@ class ImportAccountAudience extends Command
             $job = match ($account->provider) {
                 'twitter' => ImportTwitterFollowersJob::class,
                 'facebook_page' => ImportFacebookPageFollowersJob::class,
+                'instagram' => ImportInstagramFollowersJob::class,
                 'mastodon' => ImportMastodonFollowersJob::class,
                 default => null,
             };